refactor: extract pexip plugin logic to a hook
